# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Global Enterprise
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type CreateApplicationBodyService =
  (typeof CreateApplicationBodyService)[keyof typeof CreateApplicationBodyService];

export const CreateApplicationBodyService = {
  Air_Ticket_Booking: "Air Ticket Booking",
  Train_Ticket_Booking: "Train Ticket Booking",
  Bus_Ticket_Booking: "Bus Ticket Booking",
  International_Parcel_Booking: "International Parcel Booking",
  PAN_Card_Apply: "PAN Card Apply",
  Aadhaar_Card_Services: "Aadhaar Card Services",
  Voter_Card_Apply: "Voter Card Apply",
  Passport_Apply: "Passport Apply",
  Learning_License: "Learning License",
  Driving_License: "Driving License",
  UDID_Certificate_Apply: "UDID Certificate Apply",
  "E-Shram_Card": "E-Shram Card",
  Schedule_Caste_Certificate: "Schedule Caste Certificate",
  Punjab_Resident_Certificate: "Punjab Resident Certificate",
  Income_Certificate: "Income Certificate",
  "UDYAM_Certificate_(MSME)": "UDYAM Certificate (MSME)",
  GST_Registration: "GST Registration",
  "Job_Application_Forms_(Govt_Naukri)": "Job Application Forms (Govt Naukri)",
  College_Admission_Forms: "College Admission Forms",
  School_Admission_Forms: "School Admission Forms",
  Competitive_Exam_Forms: "Competitive Exam Forms",
  Scholarship_Forms: "Scholarship Forms",
  General_Online_Form_Filling: "General Online Form Filling",
  Document_Scanning: "Document Scanning",
  Printing_Services: "Printing Services",
  Website_Design_Services: "Website Design Services",
  "AEPS_(Aadhaar_Enabled_Payment_System)":
    "AEPS (Aadhaar Enabled Payment System)",
  Online_Payments: "Online Payments",
} as const;

export interface CreateApplicationBody {
  /**
   * @minLength 2
   * @maxLength 100
   */
  name: string;
  /**
   * @minLength 10
   * @maxLength 15
   */
  phone: string;
  /** @maxLength 254 */
  email?: string;
  service: CreateApplicationBodyService;
  /** @maxLength 1000 */
  message?: string;
  callbackRequested?: boolean;
  /**
   * JSON-encoded service-specific application details
   * @maxLength 12000
   */
  details?: string;
}

export interface Application {
  id: number;
  trackingNumber: string;
  name: string;
  phone: string;
  email?: string | null;
  service: string;
  message?: string | null;
  status: string;
  callbackRequested: boolean;
  paymentStatus?: string;
  paymentAmount?: number | null;
  pricingType?: string;
  pricingStatus?: string;
  applicationPrice?: number | null;
  paidAt?: string | null;
  /**
   * JSON-encoded service-specific application details
   * @maxLength 12000
   */
  details: string;
  createdAt: string;
}

export interface ApplicationReceipt {
  trackingNumber: string;
  name: string;
  phone: string;
  email?: string | null;
  service: string;
  paymentAmount: number;
  paymentTxnId: string;
  paidAt: string;
  createdAt: string;
}

export interface PayUPaymentFields {
  [key: string]: string;
}

export interface ApplicationPayment {
  required: boolean;
  action?: string;
  fields?: PayUPaymentFields;
  amount: number;
  baseAmount: number;
  gatewayFee: number;
  gatewayFeeGst: number;
}

export type ApplicationCreateResponse = Application & {
  payment?: ApplicationPayment;
};

export interface ApplicationListResponse {
  applications: Application[];
  total: number;
}

export interface TrackApplicationResponse {
  trackingNumber: string;
  service: string;
  status: string;
  callbackRequested: boolean;
  paymentStatus?: string;
  paymentAmount?: number | null;
  pricingType?: string;
  pricingStatus?: string;
  applicationPrice?: number | null;
  createdAt: string;
}

export interface SetApplicationPriceBody {
  /**
   * @minimum 0
   * @maximum 100000000
   */
  price: number;
  /** @maxLength 5000 */
  internalNotes?: string | null;
}

export interface ApplicationPricingResponse {
  id: number;
  pricingType: string;
  pricingStatus: string;
  applicationPrice: number;
  paymentAmount: number;
  paymentStatus: string;
  priceAssignedAt?: string | null;
  priceAssignedBy?: string | null;
}

export interface CreatePaymentRequestBody {
  /** @maxLength 200 */
  service: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  /** @maxLength 20 */
  phone?: string;
  email?: string | null;
  /**
   * @minimum 1
   * @maximum 100000000
   */
  amount: number;
  /** @maxLength 5000 */
  notes?: string | null;
}

export interface PaymentRequestPublic {
  token: string;
  service: string;
  name: string;
  amount: number;
  paymentStatus: string;
  paidAt?: string | null;
  createdAt: string;
}

export type AdminPaymentSource =
  (typeof AdminPaymentSource)[keyof typeof AdminPaymentSource];

export const AdminPaymentSource = {
  application: "application",
  manual: "manual",
} as const;

export interface AdminPayment {
  id: string;
  source: AdminPaymentSource;
  reference: string;
  service: string;
  clientName: string;
  email?: string | null;
  phone?: string | null;
  amount: number;
  paymentStatus: string;
  transactionId?: string | null;
  paidAt?: string | null;
  createdAt: string;
}

export interface AdminPaymentsResponse {
  payments: AdminPayment[];
}

export type PaymentRequestCreateResponse = PaymentRequestPublic & {
  paymentPageUrl: string;
};

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  token: string;
  username: string;
}

export interface ServiceCount {
  service: string;
  count: number;
}

export interface DashboardStats {
  totalApplications: number;
  recentApplications: Application[];
  applicationsByService: ServiceCount[];
  visitorCount: number;
  totalLeads?: number;
}

export interface ServicePrice {
  service: string;
  /** @minimum 0 */
  price: number;
  updatedAt: string;
}

export interface ServicePriceList {
  prices: ServicePrice[];
}

export interface VisitorCount {
  count: number;
}

export type ListApplicationsParams = {
  /**
   * Filter by service type
   */
  service?: string;
  limit?: number;
  offset?: number;
};

export type DeleteApplication200 = {
  success: boolean;
  id: number;
};

export type ExportApplicationsCsvParams = {
  service?: string;
};

export type SetServicePriceBody = {
  /** @minimum 0 */
  price: number;
};