# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Global Enterprise
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/dashboard:
    get:
      operationId: getDashboardStats
      tags: [admin]
      summary: Get dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/pricing:
    get:
      operationId: listServicePrices
      tags: [admin]
      summary: List service prices (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Service prices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicePriceList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/pricing/{service}:
    put:
      operationId: setServicePrice
      tags: [admin]
      summary: Set a service price (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: service
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                price:
                  type: integer
                  minimum: 0
              required:
                - price
      responses:
        "200":
          description: Price saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicePrice"
        "400":
          description: Invalid price
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Application:
      type: object
      properties:
        id:
          type: integer
        trackingNumber:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
          nullable: true
        service:
          type: string
        message:
          type: string
          nullable: true
        status:
          type: string
        callbackRequested:
          type: boolean
        paymentStatus:
          type: string
        paymentAmount:
          type: number
          format: double
          nullable: true
        pricingType:
          type: string
        pricingStatus:
          type: string
        applicationPrice:
          type: number
          format: double
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        details:
          type: string
          maxLength: 12000
          description: JSON-encoded service-specific application details
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - trackingNumber
        - name
        - phone
        - service
        - status
        - callbackRequested
        - details
        - createdAt
...
    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
      required:
        - token
        - username
    DashboardStats:
      type: object
      properties:
        totalApplications:
          type: integer
        recentApplications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        applicationsByService:
          type: array
          items:
            $ref: "#/components/schemas/ServiceCount"
        visitorCount:
          type: integer
        totalLeads:
          type: integer
      required:
        - totalApplications
        - recentApplications
        - applicationsByService
        - visitorCount
    ServicePrice:
      type: object
      properties:
        service:
          type: string
        price:
          type: integer
          minimum: 0
        updatedAt:
          type: string
          format: date-time
      required:
        - service
        - price
        - updatedAt
    ServicePriceList:
      type: object
      properties:
        prices:
          type: array
          items:
            $ref: "#/components/schemas/ServicePrice"
      required:
        - prices
    ServiceCount:
      type: object
      properties:
        service:
          type: string
        count:
          type: integer
      required:
        - service
        - count
...